requirements.txt
aiohttp==3.11.11
httpx==0.28.1
Jinja2==3.1.6
joblib==1.4.2
json_repair==0.34.0
langchain_community==0.3.21
litellm==1.66.0
llama_index==0.12.30
matplotlib==3.10.1
nest_asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
pandas==2.2.3
pyaml_env==1.2.2
pydantic==2.11.3
PyMySQL==1.1.1
python-dotenv==1.1.0
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.6.1
tiktoken==0.7.0
xgboost==3.0.0
zhipuai==2.1.5.20250410
